Avoid implementation-defined narrowing in correctionLaunch()
correctionLaunch() accumulated NO_FUEL_CORRECTION (100) + lnchFuelAdd
(tune-editable -40..+80) in an int8_t, which can hold a computed
value of 180 only via an implementation-defined conversion on
assignment (well-defined as two's-complement wraparound on this
project's GCC-based toolchains, but not something the standard
guarantees generally). Widened the accumulator to int16_t so the
value is correct by construction rather than by relying on a specific
narrowing-conversion behaviour; the final (uint8_t) cast is now a
plain, always-safe conversion since the result is always in [60,180].

Fixes speeduino#1571

+++ speeduino/corrections.cpp
@@ -568,10 +568,10 @@ This simple check applies the extra fuel if we're currently launching
 */
 TESTABLE_INLINE_STATIC uint8_t correctionLaunch(void)
 {
-  int8_t correction = (int8_t)NO_FUEL_CORRECTION;
+  int16_t correction = (int16_t)NO_FUEL_CORRECTION;
   if (currentStatus.launchingHard || currentStatus.launchingSoft) {
     correction = correction + configPage6.lnchFuelAdd;
-  } 
+  }
   return (uint8_t)correction;
 }
